tests/wpt/web-platform-tests/css/css-contain/container-queries/container-name-computed.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Containment Test: Computed values of container-name</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#container-name">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="support/cq-testcommon.js"></script>
<div id="target"></div>
<script>
setup(() => assert_implements_container_queries());

test_computed_value('container-name', 'initial', 'none');
test_computed_value('container-name', 'unset', 'none');
test_computed_value('container-name', 'foo');
test_computed_value('container-name', 'FoO');
test_computed_value('container-name', 'foo bar');
test_computed_value('container-name', 'bar foo');
test_computed_value('container-name', 'foo foo bar', 'foo bar');
test_computed_value('container-name', 'foo bar foo', 'foo bar');
test_computed_value('container-name', 'bar foo foo', 'bar foo');
test_computed_value('container-name', '\\!escaped');
</script>
